Add files via upload
This is a python code for comprehensive Data Visualization for a car company that uses its portal to sell used/new cars. The data used for this activity contains details of the cars in terms of mileage, colour, model name, body type, year of registration, price of the car and so on. Using python, I have cleaned the data and by using different visualization tools like heat plots, bar plots, count plots, boxplots, etc,  I have tried to find the insights as to which of the car properties are most relevant and contribute the most towards its respective price.