Add files via upload
This is a python code for the implementation of basic Supervised Machine learning models for a car company that uses its portal to sell used/new cars. The data used for this activity contains details of the cars in terms of mileage, colour, model name, body type, year of registration, price of the car and so on. 
This code consists of data cleaning, feature engineering, supervised regression models (linear regression and polynomial regression) and supervised classification models (Support Vector Machine and Decision Tree). After training and testing the models, I have also compared the predicted data with the test data (using tabular and graphical representations) to access the performance of the model by evaluating its accuracy in order to determine the best suitable model for the given data.